SVN Bug Fixed
Support for .tar.gz and .tgz extensions, or simply gz compressed tars
has been removed because  they don't let you "update" the archive
without recompressing everything.

zip let's you do this and corpora controller has been updated so
that after a corpus is checked in the original zip file is updated correctly.

another option was tried and tested:
after every upload overwriting the .svn sub-folder -- pretending that
every change was made from the most recent copy. This worked but it
essentially ignored conflicts. This method would work better if users
download history was tracked and stored in some table, and the user was forced
to always update from a recent copy. But this method was aborted after
some trial and error.

Everything _seems_ to work correctly. More testing to follow.