Completed "Given I am logged in" step for Cucumber Testing in common_steps.rb
Generates me as a test user.
Confirms users and Skips email confirmation.
Correctly validates whether or not I am logged in

This is to be used in many future tests including tests
for Corpus Form.

To-Do: Switch to Factory Girl for generating test Users.

This doesn't matter that much because... Database Cleaner.